# Conflicts:
#	.idea/workspace.xml
#	src/uet/oop/bomberman/entities/Character/Bomberman.java
#	target/classes/uet/oop/bomberman/entities/Character/Bomberman.class

+++ src/uet/oop/bomberman/entities/Character/Character.java
@@ -27,7 +27,7 @@ public abstract class Character extends Entity {
 
     public abstract void afterKill();
 
-    protected abstract boolean canMove(double x, double y);
+    protected abstract boolean canMove(int x, int y);
 
     protected abstract void calculateMove();
